вот это я понимаю рефактор
