+++ kpi_radio_bot/broadcast/__init__.py
@@ -1,0 +1,2 @@
+from . import radioboss, playlist
+from .broadcast import *
