+++ kpi_radio_bot/config.py
@@ -6,13 +6,12 @@ import aiohttp
 from aiogram import Bot
 from dotenv import load_dotenv
 
-
 PATH_SELF = Path(__file__).parent.parent
 PATH_STUFF = PATH_SELF / 'stuff'
+PATH_LOG = PATH_STUFF / 'debug.log'
 
 load_dotenv(dotenv_path=PATH_STUFF / '.env')
-logging.basicConfig(filename=PATH_STUFF / 'debug.log', level=logging.INFO)
-
+logging.basicConfig(filename=PATH_LOG, level=logging.INFO)
 
 IS_TEST_ENV = getenv("IS_TEST_ENV")
