+++ kpi_radio_bot/consts/__init__.py
@@ -1,0 +1,4 @@
+from . import btns_text
+from . import keyboards
+from . import texts
+from .others import *
