+++ kpi_radio_bot/consts/btns_text.py
@@ -1,0 +1,28 @@
+MENU = {
+    'order': '📝Заказать песню',
+    'what_playing': '🎧Что играет?',
+    'help': '⁉️Помощь',
+    'timetable': '⏱Расписание эфиров',
+    'feedback': '🖌Обратная связь',
+}
+
+HELP = {
+    'orders': '📝Заказ песни',
+    'criteria': '❗Модерация',
+    'playlist': '⏭Плейлист',
+    'feedback': '🖌Обратная связь'
+}
+
+INLINE_SEARCH = 'Удобный поиск'
+
+HISTORY = 'Что играло'
+NEXT = 'Что будет играть'
+
+BACK = 'Назад'
+CANCEL = 'Отмена'
+
+QUEUE = '✅Принять'
+NOW = 'Сейчас'
+REJECT = '❌Отклонить'
+
+BAD_ORDER_BUT_OK = 'Все ок'
