+++ kpi_radio_bot/consts/others.py
@@ -1,0 +1,82 @@
+from pathlib import Path
+
+BAD_WORDS = [
+    'пизд',
+    'бля',
+    'хуй', 'хуя', 'хуи', 'хуе',
+    'жирный член',
+    'ебать', 'еби', 'ебло', 'ебля', 'ебуч',
+    'долбо',
+    'дрочит',
+    'мудак', 'мудило',
+    'пидор', 'пидар',
+    'сука', 'суку',
+    'гандон',
+    'fuck', 'bitch', 'shit', 'dick', 'cunt'
+]
+
+ANIME_WORDS = ['anime', 'аниме']
+
+BAD_NAMES = [
+    'корж', 'тима', 'стрыкало', 'нервы', 'гречка',
+    'gachi', '♂', 'slave', 'ass', 'butt', 'right version',
+    'remix', 'radio tapok',
+]
+
+TIMES_NAME = {
+    'week_days': ['Понедельник', 'Вторник', 'Среда', 'Четверг', 'Пятница', 'Суббота', 'Воскресенье'],
+    'next_days': ['Сегодня', 'Завтра', 'Послезавтра', 'Послепослезавтра', 'Сейчас'],
+    'times': ['Утренний эфир', 'Первый перерыв', 'Второй перерыв', 'Третий перерыв', 'Четвертый перерыв',
+              'Вечерний эфир'],
+}
+
+BROADCAST_TIMES = {  # todo короче эту тему лучше сделать как функцию геттер имхо
+    #  day:
+    #       num:  start, stop
+
+    **dict.fromkeys(  # same value for many keys
+        [0, 1, 2, 3, 4, 5],
+        {
+            0: ('08:00', '08:30'),
+            1: ('10:05', '10:25'),
+            2: ('12:00', '12:20'),
+            3: ('13:55', '14:15'),
+            4: ('15:50', '16:10'),
+            5: ('17:50', '22:00'),
+        }
+    ),
+
+    6: {
+        0: ('12:00', '18:00'),
+        5: ('18:00', '22:00')
+    }
+
+}
+
+PATHS = {
+    'orders': Path('D:/Вещание Радио/Заказы'),  # сюда бот кидает заказанные песни
+    'archive': Path('D:/Вещание Радио/Архив'),  # сюда песни перемещаются каждую ночь с папки заказов
+    'ether': Path('D:/Вещание Радио/Эфир'),  # тут песни выбранные радистами, не используется
+}
+
+
+# летние пути
+# paths = {
+#     'orders': Path('D:/Вещание Радио/Летний эфир/Заказы'),
+#     'archive': Path('D:/Вещание Радио/Летний эфир/Архив'),
+#     'ether': Path('D:/Вещание Радио/Летний эфир/Эфир'),
+# }
+
+
+def _time(time_string):
+    hours, mins = time_string.split(':')
+    return int(hours) * 60 + int(mins)
+
+
+BROADCAST_TIMES_ = {
+    day_k: {
+        num_k: tuple(_time(time) for time in num_v)
+        for num_k, num_v in day_v.items()
+    }
+    for day_k, day_v in BROADCAST_TIMES.items()
+}
