+++ kpi_radio_bot/core/communication.py
@@ -1,7 +1,8 @@
 from collections import OrderedDict
 
+import utils.get_by
 from config import BOT, ADMINS_CHAT_ID
-from utils import other
+from utils import user_utils
 
 # key value db: to_message_id: (from_chat_id, from_message_id)
 MESSAGES_CACHE = OrderedDict()
@@ -28,8 +29,8 @@ async def user_message(message):
     else:
         reply_to = None
 
-    text = f"От {other.get_user_name(message.from_user)}: \n"
-    await resend_message(message, ADMINS_CHAT_ID, additional_text=text, reply_to=reply_to)
+    text = f"От {utils.get_by.get_user_name(message.from_user)}: \n"
+    await _resend_message(message, ADMINS_CHAT_ID, additional_text=text, reply_to=reply_to)
 
 
 async def admin_message(message):
@@ -40,33 +41,37 @@ async def admin_message(message):
         user, reply_to = cache_get(message.reply_to_message.message_id)
         text = ''
     else:
-        user = other.get_user_from_entity(message.reply_to_message)
+        user = user_utils.get_user_from_entity(message.reply_to_message)
         reply_to = None
         if not user:
             return
         user = user.id
 
         if message.reply_to_message.audio:
-            text = "На ваш заказ <i>(" + other.get_audio_name(message.reply_to_message.audio) + ")</i> ответили: \n"
+            text = "На ваш заказ <i>(" + utils.get_by.get_audio_name(
+                message.reply_to_message.audio) + ")</i> ответили: \n"
         else:
             text = "На ваше сообщение ответили: \n"
 
-    await resend_message(message, user, additional_text=text, reply_to=reply_to)
+    await _resend_message(message, user, additional_text=text, reply_to=reply_to)
 
 
-async def resend_message(message, chat, additional_text='', reply_to=None):
+#
+
+
+async def _resend_message(message, chat, additional_text='', reply_to=None):
     if additional_text and (message.audio or message.sticker):
-        m = await BOT.send_message(chat, additional_text)
-        cache_add(m.message_id, message)
+        mes = await BOT.send_message(chat, additional_text)
+        cache_add(mes.message_id, message)
 
     if message.audio:
-        m = await BOT.send_audio(chat, message.audio.file_id, reply_to_message_id=reply_to)
+        mes = await BOT.send_audio(chat, message.audio.file_id, reply_to_message_id=reply_to)
     elif message.sticker:
-        m = await BOT.send_sticker(chat, message.sticker.file_id, reply_to_message_id=reply_to)
+        mes = await BOT.send_sticker(chat, message.sticker.file_id, reply_to_message_id=reply_to)
     elif message.photo:
-        m = await BOT.send_photo(chat, message.photo[-1].file_id,
-                                 caption=additional_text + (message.caption or ''), reply_to_message_id=reply_to)
+        mes = await BOT.send_photo(chat, message.photo[-1].file_id,
+                                   caption=additional_text + (message.caption or ''), reply_to_message_id=reply_to)
     else:
-        m = await BOT.send_message(chat, additional_text + (message.text or ''), reply_to_message_id=reply_to)
+        mes = await BOT.send_message(chat, additional_text + (message.text or ''), reply_to_message_id=reply_to)
 
-    cache_add(m.message_id, message)
+    cache_add(mes.message_id, message)
